+++ lib/vagrant/action/vm/check_box.rb
@@ -17,6 +17,7 @@ module Vagrant
             env.ui.info I18n.t("vagrant.actions.vm.check_box.not_found", :name => box_name)
             Vagrant::Box.add(env.env, box_name, box_url)
             env["boxes"].reload!
+            env.env.reload_config!
           end
 
           @app.call(env)